[App] [Checkpoint] Checkpoint before breaking down PeakNote into PeakBook, PeakLearning, and PeakNote
