+++ peak-app/src/common/rich-text-editor/editors/scratchpad/constants.ts
@@ -1,6 +1,6 @@
-import {ScratchPad} from "../../../../constants/wiki-types";
 import {INITIAL_PAGE_STATE} from "../../../../constants/editor";
 import { Node } from "slate";
+import {ScratchPad} from "../../../../types/notes";
 
 export const SCRATCHPAD_ID = "scratchpad"
 export const SCRATCHPAD_TITLE = "scratchpad"
