+++ peak-app/src/types/editors.ts
@@ -1,0 +1,5 @@
+import {PeakWikiPage, ScratchPad} from "./notes";
+
+export interface PeakWikiState {
+    [key: string]: PeakWikiPage | ScratchPad;
+}
