+++ peak-app/src/utils/notes.ts
@@ -1,8 +1,9 @@
-import {PeakNote, upsertNote} from "../redux/slices/noteSlice";
+import {upsertNote} from "../redux/slices/noteSlice";
 import {fetchNewestNote} from "../client/notes";
 import {Peaker} from "../types";
 import {currentUserInRedux} from "../redux/utils";
 import {store} from "../redux/store";
+import {PeakNote} from "../types/notes";
 
 export function newestNodeAcrossAllAcounts(): Promise<PeakNote | null> {
     const currentUser: Peaker = currentUserInRedux()
