+++ peak-app/src/views/notes/list/note-list-view.scss
