+++ peak-app/src/views/notes/view/note-editor/peak-note-editor.scss
