+++ peak-app/src/views/notes/view/note-skeleton/NoteSkeleton.tsx
