+++ peak-app/src/views/notes/view/note-skeleton/note-skeleton.scss
