+++ peak-app/src/views/notes/view/note-view.scss
